Changes database DSNs to be configured in full
The previous pdp_config.j2 template only had a parameter for the
username for each dsn, and had one common password that was
supposed to work for both dsns. This was not general enough,
in that it didn't let you configure the hostname, database name,
or the protocol, or have a different password for each database.

Instead, revert to these variables being configurable as one entire
SQLAlchemy DSN string.