Removes two layers of indirection from deployment
Our Docker deployment setup has had multiple layers of indirection:
1) Docker
2) nginx
3) supervisord
4) gunicorn
5) pdp

nginx was just being used as a proxy to glue the services together, but
we're better off using a proxy at a higher level, so we removed this
from the setup (which just made it more complicated to set up).

Likewise supervisord was being used to keep the services running, but
it's pointless when the Docker container management is doing the same
thing. It has been removed.

With this change, the setup is substantially simplifed. Docker simply
runs (a configurable number of) gunicorn instances that run either pdp
frontend or backend. All the orchestration can now happen *outside* of
the package, where it belongs.