Adds massively expensive BCCAQ download tests
Adds a set of tests that actually query the BCCAQ catalogs, construct
a URL for every dataset and downloads a subset and checks it.
It takes forever but it's very comprehensive. At present the tests
are failing as most of the BCCAQv2 data doesn't match the content
length. More work necessary.