Reorganization of the WSGI modules to isolate
There existed a bit of friction with running either PCDS portal or a
raster portal in isolation. The pdp.wsgi module *imported* things from
both of these, so it wasn't technically possible to run one or the
other without the full requirements. This patch moves the wsgi app
functions to their own modules, so one can pick and choose one or the
other or both.

Ultimately, this is so that we can run a PCDS portal without pulling
in any requirements of modelmeta (and all of the discrete geometry
work that still isn't quite migrated in the database :/ )