Correct broken tests from PR #188
PR #188 changed some things around to accommodate ncWMS 2. In order to
make the changes some of the base data files (including a few that
were referenced in the test suite), were split up and rewritten and no
longer exist in the production app.

This changes modifies the test code to match the new data.